requirements.txt
torch>=1.12.0
numpy>=1.21.0
python-chess>=1.9.0
graphviz>=0.20.0
matplotlib>=3.5.0